+++ app/src/main/java/com/example/myprojectforandroidacademy/FragmentMoviesDetails.kt
@@ -1,17 +1,12 @@
 package com.example.myprojectforandroidacademy
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 
-class FragmentMoviesDetailsActivity : Fragment() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        //setContentView(R.layout.fragment_movies_details)
-    }
+class FragmentMoviesDetails : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
