--- app/src/main/res/layout-land/fragment_movies_list.xml
@@ -1,161 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/myBlack"
-    tools:context=".FragmentMoviesListActivity">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/moviesList"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/margin16dp"
-            android:layout_marginTop="@dimen/margin30dp"
-            android:drawableStart="@drawable/ic_combined_shape"
-            android:text="@string/movies_list"
-            android:textColor="@color/white"
-            android:textSize="@dimen/text_size14"
-            android:textStyle="bold"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.cardview.widget.CardView
-            android:id="@+id/cardView"
-            android:layout_width="wrap_content"
-            android:layout_height="300dp"
-            android:layout_marginStart="@dimen/margin14dp"
-            android:layout_marginTop="@dimen/margin18dp"
-            android:background="@color/black_gray"
-            app:cardCornerRadius="@dimen/corner10dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/moviesList">
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:background="@color/black_gray">
-
-                <ImageView
-                    android:id="@+id/movieImage"
-                    android:layout_width="165dp"
-                    android:layout_height="250dp"
-                    android:scaleType="centerCrop"
-                    android:src="@drawable/movie"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <ImageView
-                    android:id="@+id/movieImageGradient"
-                    android:layout_width="165dp"
-                    android:layout_height="250dp"
-                    android:scaleType="centerCrop"
-                    android:src="@drawable/movie_gradient_list"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-
-                <TextView
-                    android:id="@+id/movieAge"
-                    android:layout_width="24dp"
-                    android:layout_height="24dp"
-                    android:layout_marginStart="@dimen/margin8dp"
-                    android:layout_marginTop="@dimen/margin6dp"
-                    android:background="@drawable/text_view_corner"
-                    android:gravity="center"
-                    android:text="@string/thirteen_years"
-                    android:textColor="@color/white"
-                    android:textSize="@dimen/text_size12"
-                    android:textStyle="bold"
-                    app:layout_constraintStart_toStartOf="@id/movieImage"
-                    app:layout_constraintTop_toTopOf="@id/movieImage" />
-
-                <ToggleButton
-                    android:layout_width="16dp"
-                    android:layout_height="14dp"
-                    android:layout_marginTop="@dimen/margin13dp"
-                    android:layout_marginEnd="@dimen/margin6dp"
-                    android:background="@drawable/ic_like"
-                    android:text=""
-                    android:textOff=""
-                    android:textOn=""
-                    app:layout_constraintEnd_toEndOf="@id/movieImage"
-                    app:layout_constraintTop_toTopOf="@id/movieImage" />
-
-                <TextView
-                    android:id="@+id/categoriesMovie"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-
-                    android:layout_marginTop="@dimen/margin4dp"
-                    android:text="@string/categories_movie"
-                    android:textColor="@color/myRed"
-                    android:textSize="@dimen/text_size8"
-                    app:layout_constraintBottom_toTopOf="@id/ratingMovie"
-                    app:layout_constraintStart_toStartOf="@id/ratingMovie" />
-
-                <RatingBar
-                    android:id="@+id/ratingMovie"
-                    style="@style/Widget.AppCompat.RatingBar.Small"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/margin6dp"
-                    android:layout_marginBottom="@dimen/margin10dp"
-                    android:numStars="5"
-                    android:progressBackgroundTint="@color/gray"
-                    android:progressTint="@color/pink"
-                    android:rating="4"
-                    android:secondaryProgressTint="@color/pink"
-                    android:stepSize="0.5"
-                    app:layout_constraintBottom_toBottomOf="@id/movieImage"
-                    app:layout_constraintStart_toStartOf="@+id/movieImage" />
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/margin8dp"
-                    android:text="@string/count_reviews"
-                    android:textColor="@color/gray2"
-                    android:textSize="@dimen/text_size8"
-                    android:textStyle="bold"
-                    app:layout_constraintBottom_toBottomOf="@id/ratingMovie"
-                    app:layout_constraintStart_toEndOf="@id/ratingMovie"
-                    app:layout_constraintTop_toTopOf="@id/ratingMovie" />
-
-                <TextView
-                    android:id="@+id/movieTitle"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/margin8dp"
-                    android:layout_marginTop="@dimen/margin8dp"
-                    android:text="@string/title_movie"
-                    android:textColor="@color/white"
-                    android:textSize="@dimen/text_size14"
-                    android:textStyle="bold"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/movieImage" />
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/margin8dp"
-                    android:text="@string/duration_film"
-                    android:textColor="@color/gray2"
-                    android:textSize="@dimen/text_size8"
-                    android:textStyle="bold"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/movieTitle" />
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-        </androidx.cardview.widget.CardView>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</ScrollView>
