+++ app/src/main/res/layout/fragment_movies_details.xml
@@ -1,14 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/myBlack">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content">
 
         <ImageView
             android:id="@+id/movieImage"
@@ -16,6 +15,7 @@
             android:layout_height="wrap_content"
             android:scaleType="centerCrop"
             android:src="@drawable/background_image"
+            android:contentDescription="@string/movie_poster"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
@@ -26,11 +26,23 @@
             android:layout_height="wrap_content"
             android:scaleType="centerCrop"
             android:src="@drawable/movie_gradient"
+            android:contentDescription="@string/movie_poster"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="@dimen/text_size12"
+            android:text="@string/back"
+            android:drawableStart="@drawable/less"
+            android:drawablePadding="@dimen/margin8dp"
+            app:layout_constraintTop_toTopOf="@id/movieImage"
+            app:layout_constraintStart_toStartOf="@id/movieImage"
+            android:layout_marginStart="@dimen/margin22dp"
+            android:layout_marginTop="@dimen/margin56dp"
+            android:textColor="@color/gray"/>
 
         <TextView
             android:layout_width="wrap_content"
@@ -142,6 +154,7 @@
             android:layout_marginTop="@dimen/margin7dp"
             android:scaleType="centerCrop"
             android:src="@drawable/cast_robert_downey"
+            android:contentDescription="@string/cast_robert_downey"
             app:layout_constraintBottom_toTopOf="@+id/castRobertDowney"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintEnd_toStartOf="@+id/imageCastChrisEvams"
@@ -156,6 +169,7 @@
             android:layout_marginTop="@dimen/margin7dp"
             android:scaleType="centerCrop"
             android:src="@drawable/cast_chris_evans"
+            android:contentDescription="@string/cast_chris_evans"
             app:layout_constraintBottom_toTopOf="@id/castChrisEvans"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintEnd_toStartOf="@+id/imageCastMarkRufallo"
@@ -170,6 +184,7 @@
             android:layout_marginTop="@dimen/margin7dp"
             android:scaleType="centerCrop"
             android:src="@drawable/cast_mark_ruffalo"
+            android:contentDescription="@string/cast_mark_ruffalo"
             app:layout_constraintBottom_toTopOf="@id/castMarkRufallo"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintEnd_toStartOf="@+id/imageCastChrisHemsworth"
@@ -185,6 +200,7 @@
             android:layout_marginEnd="@dimen/margin16dp"
             android:scaleType="centerCrop"
             android:src="@drawable/cast_chris_hemsworth"
+            android:contentDescription="@string/cast_chris_hemsworth"
             app:layout_constraintBottom_toTopOf="@id/castChrisHemsworth"
             app:layout_constraintDimensionRatio="1:1"
             app:layout_constraintEnd_toEndOf="parent"
