+++ app/src/main/res/layout/fragment_movies_list.xml
@@ -5,19 +5,20 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/myBlack"
-    tools:context=".FragmentMoviesListActivity">
+    tools:context=".FragmentMoviesList">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content">
 
         <TextView
             android:id="@+id/moviesList"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="@dimen/margin16dp"
-            android:layout_marginTop="@dimen/margin55dp"
-            android:drawableLeft="@drawable/ic_combined_shape"
+            android:layout_marginTop="@dimen/margin44dp"
+            android:drawableStart="@drawable/ic_combined_shape"
+            android:drawablePadding="@dimen/margin8dp"
             android:text="@string/movies_list"
             android:textColor="@color/white"
             android:textSize="@dimen/text_size14"
@@ -47,6 +48,7 @@
                     android:layout_height="250dp"
                     android:scaleType="centerCrop"
                     android:src="@drawable/movie"
+                    android:contentDescription="@string/movie_poster"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
@@ -69,7 +71,7 @@
                     android:layout_marginTop="@dimen/margin6dp"
                     android:background="@drawable/text_view_corner"
                     android:gravity="center"
-                    android:text="13+"
+                    android:text="@string/thirteen_years"
                     android:textColor="@color/white"
                     android:textSize="@dimen/text_size12"
                     android:textStyle="bold"
@@ -81,7 +83,7 @@
                     android:layout_height="14dp"
                     android:layout_marginTop="@dimen/margin13dp"
                     android:layout_marginEnd="@dimen/margin6dp"
-                    android:background="@drawable/ic_like"
+                    android:background="@drawable/ic_like_white"
                     android:text=""
                     android:textOff=""
                     android:textOn=""
@@ -134,7 +136,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="@dimen/margin8dp"
                     android:layout_marginTop="@dimen/margin8dp"
-                    android:text="Avengers: End Game"
+                    android:text="@string/title_movie"
                     android:textColor="@color/white"
                     android:textSize="@dimen/text_size14"
                     android:textStyle="bold"
@@ -145,10 +147,11 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="@dimen/margin8dp"
-                    android:text="137 MIN"
+                    android:text="@string/duration_film"
                     android:textColor="@color/gray2"
                     android:textSize="@dimen/text_size8"
                     android:textStyle="bold"
+                    android:textAllCaps="true"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/movieTitle" />
